[FE][Fix][Refactor] : HeaderDropdownProvider.tsx -> HeaderDropdownContext.tsx 로 수정 및 디렉토리 이동
